Rename manager -> dashboard and reorder components for it
