+++ src/app/dashboard/page.tsx
@@ -11,21 +11,21 @@ const yourAccount = [
 		id: 1,
 		image: "/images/icons/calendar-icon.svg",
 		imageAlt: "Calendar icon",
-		link: "/manager/subscription-managment/",
+		link: "/dashboard/subscription-managment/",
 		title: "Manage Subscriptions",
 	},
 	{
 		id: 2,
 		image: "/images/icons/clock-icon.svg",
 		imageAlt: "Clock icon",
-		link: "/manager/view-past-orders/",
+		link: "/dashboard/view-past-orders/",
 		title: "View Past Orders",
 	},
 	{
 		id: 3,
 		image: "/images/icons/user-profile-icon.svg",
 		imageAlt: "User profile icon",
-		link: "/manager/account",
+		link: "/dashboard/account",
 		title: "Manage Account",
 	},
 ] as IManagerOptionCardModel[];
@@ -35,14 +35,14 @@ const otherActions = [
 		id: 1,
 		image: "/images/icons/discount-icon.svg",
 		imageAlt: "Discount icon",
-		link: "/manager/refer/",
+		link: "/dashboard/refer/",
 		title: "Refer a Friend & You BOTH Save",
 	},
 	{
 		id: 2,
 		image: "/images/icons/cart-icon.svg",
 		imageAlt: "Cart icon",
-		link: "/manager/shop-supplement-products/",
+		link: "/dashboard/shop-supplement-products/",
 		title: "Shop Supplement Products",
 	},
 ] as IManagerOptionCardModel[];
