+++ src/components/dashboard/ShowTextBasedOnRoute.tsx
@@ -9,19 +9,19 @@ const ShowTextBasedOnRoute = () => {
 
 	const getTextForRoute = () => {
 		switch (pathname) {
-			case "/manager/manage-plans":
+			case "/dashboard/manage-plans":
 				return "Manage Plans";
-			case "/manager/top-up-checkout":
+			case "/dashboard/top-up-checkout":
 				return "Order Top Up Capsules";
-			case "/manager/subscription-managment":
+			case "/dashboard/subscription-managment":
 				return "Subscription Management";
-			case "/manager/view-past-orders":
+			case "/dashboard/view-past-orders":
 				return "View Past Orders";
-			case "/manager/refer":
+			case "/dashboard/refer":
 				return "Refer a Friend & You BOTH Save";
-			case "/manager/shop-supplement-products":
+			case "/dashboard/shop-supplement-products":
 				return "Shop Supplement Products";
-			case "/manager/account":
+			case "/dashboard/account":
 				return "Manage Account";
 			default:
 				return null;
@@ -38,7 +38,7 @@ const ShowTextBasedOnRoute = () => {
 		<div className="text-[16px] leading-[18px] md:text-[24px] md:leading-[27px] font-[400] font-hagerman flex items-center justify-center h-[62px] border-b-[1px] border-grey30 relative">
 			<Link
 				className="absolute top-1/2 left-[16px] md:left-[32px] transform -translate-y-1/2 flex items-center gap-[8px] text-blue font-bold  text-[16px] leading-[18px] font-helvetica"
-				href="/manager"
+				href="/dashboard/"
 			>
 				<ChevronLeft className="text-blue w-[20px] h-[20px]" /> Back
 			</Link>
